feat: コンポーネントをわけるようにした
